Add files via upload
Contains the KiCad capture files, PCB design files, footprint and symbol library files, and gerber files which were sent to JLCPCB in order to fabricate the board. The board was ordered and realized, and works as intended.

NOTE: 
If one is to copy this design, a resistor must be added in series with the TVS diode in order not to burn the diode once the breakdown voltage has been reached.